+++ src/main/java/com/project/applicationsocial/model/entity/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.project.applicationsocial.entity;
+package com.project.applicationsocial.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
