+++ src/main/java/com/project/applicationsocial/payload/request/RegisterRequest.java
@@ -20,8 +20,6 @@ public class RegisterRequest {
 
     private String lastName;
 
-    private String roles;
-
     private Boolean gender;
 
     private String phoneNumber;
