+++ src/main/java/com/project/applicationsocial/service/UserService.java
@@ -1,8 +1,8 @@
 package com.project.applicationsocial.service;
 
-import com.project.applicationsocial.DTO.UserDTO;
-import com.project.applicationsocial.entity.Users;
-import com.project.applicationsocial.mapper.UserMapper;
+import com.project.applicationsocial.model.DTO.UserDTO;
+import com.project.applicationsocial.model.entity.Users;
+import com.project.applicationsocial.model.mapper.UserMapper;
 import com.project.applicationsocial.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -34,26 +34,6 @@ public class UserService implements UserDetailsService {
                 .orElseThrow(() -> new UsernameNotFoundException("User not found " + username));
     }
 
-    public String addUser(Users user) {
-        if(user.getRoles() == null) {
-            user.setRoles("USER");
-        }
-        if (user.getEnable() == null) {
-            user.setEnable(true);
-        }
-        user.setPassword(encoder.encode(user.getPassword()));
-        repository.save(user);
-        return "User Added Successfully";
-    }
-
-    public UserDTO getUserById(UUID id) {
-
-        Users user = repository.findById(id).orElseThrow(() -> new UsernameNotFoundException("Not found user with id = " + id));
-
-        return UserMapper.toUserDTO(user);
-    }
-
-
     public List<UserDTO> findAllUser() {
         return repository.findAll()
                 .stream()
@@ -61,14 +41,14 @@ public class UserService implements UserDetailsService {
                 .toList();
     }
 
-//
-//    public UserDTO update(UUID id, UserDTO user) {
-//        Optional<Users> user1 = repository.findById(id);
-//        if(user1.isPresent()){
-//            return repository.save(user);
-//        }
-//        return null;
-//    }
+
+    public Users update(UUID id, Users user) {
+        Optional<Users> user1 = repository.findById(id);
+        if(user1.isPresent()){
+            return repository.save(user);
+        }
+        return null;
+    }
 
 
 }
